chore: docs and fmt
